pyproject.toml
[project]
name = "text-to-sql-agent"
version = "0.1.0"
description = ""
dependencies = [
    "google-adk>=1.18.0,<2.0.0",
    "sqlglot>=27.29.0",
    "google-genai>=1.49.0,<2.0.0",
    "python-dotenv>=1.2.1,<2.0.0",
    "psycopg2-binary>=2.9.9,<3.0.0",
]

requires-python = ">=3.10,<3.15"

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2,<9.0.0",
    "pytest-asyncio>=1.2.0,<2.0.0",
    "pytest-mock>=3.14.0,<4.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "ruff>=0.14.4,<1.0.0",
]
lint = [
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901", "B006"] # ignore line too long, too complex

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]

exclude = [".venv"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb"

[tool.pytest.ini_options]
pythonpath = "src"
filterwarnings = [
    "ignore:deprecated:DeprecationWarning:google.adk.runners",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
